feat: set max gRPC message size based on server's /meta information

+++ src/main/java/io/weaviate/client6/v1/api/InstanceMetadata.java
@@ -8,5 +8,5 @@ public record InstanceMetadata(
     @SerializedName("hostname") String hostName,
     @SerializedName("version") String version,
     @SerializedName("modules") Map<String, Object> modules,
-    @SerializedName("grpcMaxMessageSize") Long grpcMaxMessageSize) {
+    @SerializedName("grpcMaxMessageSize") Integer grpcMaxMessageSize) {
 }
